[dagster-airlift][sensor-pluggability] rename file
